Add GetProductById
Fixes #16

+++ PriskollenServer.Library/Contracts/ProductResponse.cs
@@ -6,8 +6,8 @@ public class ProductResponse
     public string Name { get; set; } = string.Empty;
     public string Brand { get; set; } = string.Empty;
     public string Image { get; set; } = string.Empty;
-    public double Volume { get; set; }
-    public double? VolumeWithLiquid { get; set; } // If applicable, if not this is null
+    public decimal Volume { get; set; }
+    public decimal? VolumeWithLiquid { get; set; } // If applicable, if not this is null
     public string VolumeUnit { get; set; } = string.Empty;
     public DateTime Created { get; set; }
     public DateTime Modified { get; set; }
